Move function for running rate engine to be a function rateengine and remove standalone function
